fixed a long-standing API wart regarding the "backing up" behavior.
prior to this commit, the raw incremental interface cannot refer to
the problematic byte not in the current input. this normally affects
the strict incremental usages only, but sometimes (most evidently,
in ISO 2022-JP) it also affects the one-shot usages.

this commit allows `upto` field in the `CodecError` struct to be
negative, and updates the existing definition of `Encoding::{decode,
encode}` to handle this situation correctly.

+++ src/encoding/codec/ascii.rs
@@ -42,7 +42,7 @@ impl Encoder for ASCIIEncoder {
                 output.write_bytes(input.as_bytes().slice_to(first_error));
                 let str::CharRange {ch: _, next} = input.char_range_at(first_error);
                 (first_error, Some(CodecError {
-                    upto: next, cause: "unrepresentable character".into_maybe_owned()
+                    upto: next as int, cause: "unrepresentable character".into_maybe_owned()
                 }))
             }
             None => {
@@ -80,7 +80,7 @@ impl Decoder for ASCIIDecoder {
             Some(first_error) => {
                 write_ascii_bytes(output, input.slice_to(first_error));
                 (first_error, Some(CodecError {
-                    upto: first_error+1, cause: "invalid sequence".into_maybe_owned()
+                    upto: first_error as int + 1, cause: "invalid sequence".into_maybe_owned()
                 }))
             }
             None => {
