+++ src/encoding/codec/utf_8.rs
@@ -177,7 +177,7 @@ impl Decoder for UTF8Decoder {
                 self.queuelen = 0;
                 write_bytes(output, input.slice(0, processed));
                 return (processed, Some(CodecError {
-                    upto: upto, cause: "invalid sequence".into_maybe_owned()
+                    upto: upto as int, cause: "invalid sequence".into_maybe_owned()
                 }));
             }
         }
